# api-spec.yaml
openapi: 3.0.3
info:
  title: Next Loyalty Program API
  description: 'Unified Loyalty Program API for Next plc omnichannel retail.

    Supports customer registration, points earning/redemption, tier management,

    notifications, staff operations, and admin configuration.

    '
  version: 1.0.0
  contact:
    name: Next Loyalty Engineering
    email: loyalty-engineering@next.co.uk
servers:
- url: https://loyalty.next.co.uk/api/v1
  description: Production
- url: https://staging.loyalty.next.co.uk/api/v1
  description: Staging
- url: http://localhost:3000/api/v1
  description: Local Development
security:
- bearerAuth: []
...
paths:
  /auth/register:
    post:
      operationId: registerCustomer
      summary: Register a new customer
      description: Create a new loyalty account. Triggers OTP verification email.
        Rate limited to 5 attempts per IP per hour.
      tags:
      - Auth
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - email
              - phone
              - password
              - termsAndConditionsVersion
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: Jane Smith
                email:
                  type: string
                  format: email
                  example: jane.smith@example.com
                phone:
                  type: string
                  pattern: ^\+[1-9]\d{1,14}$
                  example: '+447700900123'
                password:
                  type: string
                  minLength: 12
                  example: SecureP@ss2026!
                termsAndConditionsVersion:
                  type: string
                  example: '2.1'
      responses:
        '201':
          description: Customer created, OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  customerId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum:
                    - pending_verification
                  message:
                    type: string
                    example: Verification code sent to your email
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      summary: Verify OTP code
      description: Validate the 6-digit OTP sent during registration or email/phone
        change. Max 5 attempts before 30-minute lockout.
      tags:
      - Auth
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - customerId
              - otpCode
              properties:
                customerId:
                  type: string
                  format: uuid
                otpCode:
                  type: string
                  pattern: ^\d{6}$
                  example: '123456'
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                    example: true
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
                    example: 900
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/resend-otp:
    post:
      operationId: resendOtp
      summary: Resend OTP code
      description: Resend verification OTP. Rate limited to 3 requests per 10 minutes.
        60-second cooldown between resends.
      tags:
      - Auth
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - customerId
              properties:
                customerId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: OTP resent
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent:
                    type: boolean
                    example: true
                  nextResendAt:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      operationId: loginCustomer
      summary: Login with email and password
      description: Authenticate with email/password. Account locked after 5 consecutive
        failures for 30 minutes.
      tags:
      - Auth
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login/social:
    post:
      operationId: socialLogin
      summary: Login with social provider
      description: Authenticate via Google or Apple OAuth2/OIDC. Creates account if
        new user.
      tags:
      - Auth
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - provider
              - idToken
              properties:
                provider:
                  type: string
                  enum:
                  - google
                  - apple
                idToken:
                  type: string
      responses:
        '200':
          description: Social login successful
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/AuthTokenResponse'
                - type: object
                  properties:
                    isNewAccount:
                      type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      description: Exchange a valid refresh token for a new access token. Token rotation
        applied.
      tags:
      - Auth
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      operationId: logoutCustomer
      summary: Logout and revoke tokens
      description: Revoke the current refresh token and invalidate the session.
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '204':
          description: Logged out successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/reset-password/request:
    post:
      operationId: requestPasswordReset
      summary: Request password reset
      description: Send a password reset link to the registered email. Link expires
        in 1 hour. Response is always 200 to prevent email enumeration.
      tags:
      - Auth
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset link sent if account exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: If an account exists with this email, a reset link has
                      been sent
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/reset-password/confirm:
    post:
      operationId: confirmPasswordReset
      summary: Confirm password reset
      description: Set a new password using the reset token from the email link.
      tags:
      - Auth
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - token
              - newPassword
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 12
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /staff/auth/login:
    post:
      operationId: staffLogin
      summary: Staff authentication
      description: Authenticate store staff via employee ID and password. Issues short-lived
        token (4-hour expiry). MFA required for manager/admin roles.
      tags:
      - Staff
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - employeeId
              - password
              properties:
                employeeId:
                  type: string
                  example: EMP-12345
                password:
                  type: string
                terminalId:
                  type: string
                  description: POS terminal identifier for audit
                  example: POS-STORE42-01
      responses:
        '200':
          description: Staff login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
                    example: 14400
                  role:
                    type: string
                    enum:
                    - loyalty_staff
                    - loyalty_manager
                    - loyalty_admin
                  requiresMfa:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15-minute expiry)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for POS terminals and e-commerce integrations
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Unique key to ensure idempotent processing of write operations
...
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: https://loyalty.next.co.uk/problems/validation-error
        title:
          type: string
          description: Short human-readable summary
          example: Validation Error
        status:
          type: integer
          description: HTTP status code
          example: 422
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: The 'email' field must be a valid email address
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
          example: /api/v1/auth/register
        traceId:
          type: string
          description: Correlation ID for distributed tracing
          example: abc123def456
        errors:
          type: array
          description: Detailed field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Must be a valid email address
              code:
                type: string
                example: INVALID_FORMAT
      example:
        type: https://loyalty.next.co.uk/problems/validation-error
        title: Validation Error
        status: 422
        detail: One or more fields failed validation
        instance: /api/v1/auth/register
        traceId: abc123def456
        errors:
        - field: email
          message: Must be a valid email address
          code: INVALID_FORMAT
    AuthTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Access token expiry in seconds
          example: 900
        tokenType:
          type: string
          example: Bearer
  responses:
    BadRequest:
      description: Bad Request — malformed request syntax or invalid parameters
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/bad-request
            title: Bad Request
            status: 400
            detail: Request body is malformed or missing required fields
            traceId: trace-001
    Unauthorized:
      description: Unauthorized — missing or invalid authentication credentials
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/unauthorized
            title: Unauthorized
            status: 401
            detail: Access token is missing, expired, or invalid
            traceId: trace-002
    Forbidden:
      description: Forbidden — authenticated but insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/forbidden
            title: Forbidden
            status: 403
            detail: You do not have permission to perform this action
            traceId: trace-003
    NotFound:
      description: Not Found — requested resource does not exist
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found
            traceId: trace-004
    UnprocessableEntity:
      description: Unprocessable Entity — request understood but validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/validation-error
            title: Validation Error
            status: 422
            detail: One or more fields failed validation
            traceId: trace-005
            errors:
            - field: email
              message: Email already registered
              code: DUPLICATE
    InternalServerError:
      description: Internal Server Error — unexpected server-side failure
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/internal-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred. Please try again later.
            traceId: trace-006